week_2/requirements.txt
groq
openai
anthropic
streamlit
python-dotenv